txn: Split up the command module so there is one submodule per command (#8228)
