+++ src/storage/txn/commands/acquire_pessimistic_lock.rs
@@ -1,0 +1,50 @@
+// Copyright 2020 TiKV Project Authors. Licensed under Apache-2.0.
+
+use crate::storage::lock_manager::WaitTimeout;
+use crate::storage::txn::commands::{Command, CommandExt, TypedCommand};
+use crate::storage::PessimisticLockRes;
+use crate::storage::Result;
+use txn_types::{Key, TimeStamp};
+
+command! {
+    /// Acquire a Pessimistic lock on the keys.
+    ///
+    /// This can be rolled back with a [`PessimisticRollback`](Command::PessimisticRollback) command.
+    AcquirePessimisticLock:
+        cmd_ty => Result<PessimisticLockRes>,
+        display => "kv::command::acquirepessimisticlock keys({}) @ {} {} | {:?}", (keys.len, start_ts, for_update_ts, ctx),
+        content => {
+            /// The set of keys to lock.
+            keys: Vec<(Key, bool)>,
+            /// The primary lock. Secondary locks (from `keys`) will refer to the primary lock.
+            primary: Vec<u8>,
+            /// The transaction timestamp.
+            start_ts: TimeStamp,
+            lock_ttl: u64,
+            is_first_lock: bool,
+            for_update_ts: TimeStamp,
+            /// Time to wait for lock released in milliseconds when encountering locks.
+            wait_timeout: Option<WaitTimeout>,
+            /// If it is true, TiKV will return values of the keys if no error, so TiDB can cache the values for
+            /// later read in the same transaction.
+            return_values: bool,
+            min_commit_ts: TimeStamp,
+        }
+}
+
+impl CommandExt for AcquirePessimisticLock {
+    ctx!();
+    tag!(acquire_pessimistic_lock);
+    ts!(start_ts);
+    command_method!(requires_pessimistic_txn, bool, true);
+    command_method!(can_be_pipelined, bool, true);
+
+    fn write_bytes(&self) -> usize {
+        self.keys
+            .iter()
+            .map(|(key, _)| key.as_encoded().len())
+            .sum()
+    }
+
+    gen_lock!(keys: multiple(|x| &x.0));
+}
