+++ src/storage/txn/commands/commit.rs
@@ -1,0 +1,30 @@
+// Copyright 2020 TiKV Project Authors. Licensed under Apache-2.0.
+
+use crate::storage::txn::commands::{Command, CommandExt, TypedCommand};
+use crate::storage::TxnStatus;
+use txn_types::Key;
+
+command! {
+    /// Commit the transaction that started at `lock_ts`.
+    ///
+    /// This should be following a [`Prewrite`](Command::Prewrite).
+    Commit:
+        cmd_ty => TxnStatus,
+        display => "kv::command::commit {} {} -> {} | {:?}", (keys.len, lock_ts, commit_ts, ctx),
+        content => {
+            /// The keys affected.
+            keys: Vec<Key>,
+            /// The lock timestamp.
+            lock_ts: txn_types::TimeStamp,
+            /// The commit timestamp.
+            commit_ts: txn_types::TimeStamp,
+        }
+}
+
+impl CommandExt for Commit {
+    ctx!();
+    tag!(commit);
+    ts!(commit_ts);
+    write_bytes!(keys: multiple);
+    gen_lock!(keys: multiple);
+}
