+++ src/storage/txn/commands/resolve_lock_lite.rs
@@ -1,0 +1,28 @@
+// Copyright 2020 TiKV Project Authors. Licensed under Apache-2.0.
+
+use crate::storage::txn::commands::{Command, CommandExt, TypedCommand};
+use txn_types::{Key, TimeStamp};
+
+command! {
+    /// Resolve locks on `resolve_keys` according to `start_ts` and `commit_ts`.
+    ResolveLockLite:
+        cmd_ty => (),
+        display => "kv::resolve_lock_lite", (),
+        content => {
+            /// The transaction timestamp.
+            start_ts: TimeStamp,
+            /// The transaction commit timestamp.
+            commit_ts: TimeStamp,
+            /// The keys to resolve.
+            resolve_keys: Vec<Key>,
+        }
+}
+
+impl CommandExt for ResolveLockLite {
+    ctx!();
+    tag!(resolve_lock_lite);
+    ts!(start_ts);
+    command_method!(is_sys_cmd, bool, true);
+    write_bytes!(resolve_keys: multiple);
+    gen_lock!(resolve_keys: multiple);
+}
